requirements.txt
# PyTorch and related packages
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0

# Hugging Face packages
transformers>=4.21.0
datasets>=2.4.0
evaluate>=0.3.0

# Data processing and visualization
scikit-learn>=1.1.0
pandas>=1.4.0
pillow>=9.2.0
matplotlib>=3.5.0
tqdm>=4.64.0

# Optional: For advanced metrics
rouge_score>=0.1.2

# Configuration management
pyyaml>=6.0.0
